retirada de métodos não utilizados e comentários e refatoração
