+++ README.md
@@ -1,16 +1,19 @@
 # Controle de Abastecimentos
 Este app tem por objetivo o cumprimento da disciplina de Frameworks Web com Angular 2+.
-Este app deve ajudar o usuário a controlar os abastecimentos do veículo, valores gastos, postos, combustíveis utilizados e rendimento.
+Este app deve ajudar o usuário a controlar os abastecimentos do veículo, valores gastos, postos e combustíveis utilizados.
 
 ## *Protótipo*
 https://www.figma.com/file/7O5lBuhK53bysaq7wx1Nos/SK_ControleAbastecimento
 
+## *Hospedagem*
+https://moreira89.github.io/controle-combustivel/
+
 ## *Checklist*
 - [X] Criar o repositório no GitHub com a estrutura do Gitflow, ou seja, branches main e develop.
-- [X] Usar componentes de algum framework CSS (Bootstrap, Materialize ou outro).
+- [ ] Usar componentes de algum framework CSS (Bootstrap, Materialize ou outro).
 - [X] Criar o layout da aplicação com componentes, ou seja, o cabeçalho e rodapé precisam ser componentes.
 - [X] Usar pelo menos dois tipos de data-binding (Interpolation, Property Binding, Event Binding e Two Way Data Binding).
-- [X] Apresentar as telas com layout responsivo usando ou não algum framework CSS.
+- [ ] Apresentar as telas com layout responsivo usando ou não algum framework CSS.
 - [X] Construir páginas web com o conceito de componentes. 
 - [X] Passar dados via hierarquia de componentes, ou seja, usando @Input ou @Output.
 - [X] Mapear componentes à rotas no módulo de rotas.
