--- src/app/abastecimento/abastecimento.module.ts
@@ -1,15 +1,0 @@
-/*import { NgModule } from '@angular/core';
-import { HttpClientModule } from '@angular/common/http';
-import { CommonModule } from '@angular/common';
-
-import { AbastecimentoComponent } from './abastecimento.component';
-import { AbastecimentosComponent } from '../abastecimentos/abastecimentos.component';
-
-@NgModule
-({
-  declarations: [AbastecimentoComponent, AbastecimentosComponent],
-  imports: [CommonModule]
-})
-
-export class AbastecimentoModule{}
-*/
