+++ src/app/componentes/cabecalho/cabecalho.component.ts
@@ -1,5 +1,4 @@
-import { AfterViewInit, Component, OnInit } from '@angular/core';
-import { bindCallback } from 'rxjs';
+import { Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'app-cabecalho',
