+++ src/app/componentes/errors/not-found/not-found.component.css
@@ -1,0 +1,3 @@
+.titulo{
+  margin-top: 20px;
+}
