+++ src/app/componentes/errors/not-found/not-found.component.html
@@ -1,0 +1,6 @@
+<div class="container text-center">
+  <h2 class="titulo">Página inexistente.</h2>
+  <p>
+      Esta página não existe ou foi removida. Clique <a href="">aqui</a> para ir para a página principal.
+  </p>
+</div>
