+++ src/app/componentes/rodape/rodape.component.html
@@ -1,19 +1,18 @@
-<div class="container blue-grey darken-2">
+<div class="container darken-2">
     <footer class="page-footer blue-grey darken-2">
     <div class="row">
       <div class="col l6 s12">
         <h5 class="white-text">Rafael Moreira</h5>
-        <!--<p class="grey-text text-lighten-4">You can use rows and columns here to organize your footer content.</p>-->
       </div>
       <div class="col l4 offset-l2 s12">
         <ul>
           <li>
-            <a class="rodape__link" href="https://br.linkedin.com/in/rafael-moreira-gomes" target="_blank">
+            <a class="white-text" href="https://br.linkedin.com/in/rafael-moreira-gomes" target="_blank">
                 <i>LinkedIn</i>
             </a>
         </li>
         <li>
-            <a class="rodape__link" href="https://github.com/moreira89" target="_blank">
+            <a class="white-text" href="https://github.com/moreira89" target="_blank">
                 <i>GitHub</i>
             </a>
         </li>
@@ -21,10 +20,9 @@
       </div>
     </div>
 
-    <div class="footer-copyright blue-grey darken-2">
+    <div class="footer-copyright darken-2">
       <div class="container">
       © {{this.year}} Copyright
-      <!--<a class="grey-text text-lighten-4 right" href="#!">More Links</a>-->
       </div>
     </div>
   </footer>
