+++ src/app/componentes/sobre/sobre.component.ts
