+++ src/app/componentes/veiculo/veiculos/veiculos.component.ts
@@ -1,8 +1,8 @@
 import { Router, ActivatedRoute } from '@angular/router';
 import { Component, Input,  OnInit } from '@angular/core';
 
-import { Veiculo } from '../veiculo';
-import { VeiculoService } from '../veiculo.service';
+import { Veiculo } from '../../../model/veiculo';
+import { VeiculoService } from '../../../service/veiculo.service';
 
 @Component({
   selector: 'app-veiculos',
@@ -20,12 +20,9 @@ export class VeiculosComponent implements OnInit{
   }
 
   ngOnInit(){
-    console.log(this.router.url);
-    console.log(this.router.url.search("redirectAbast") > 0);
-    console.log(this.activatedRoute.snapshot.paramMap.get('veiculoId'));
     if (this.router.url.search("redirectAbast") > 0){
-      let codigo = this.activatedRoute.snapshot.paramMap.get('veiculoId');
-      this.router.navigate([`veiculo/${codigo}/abastecimentos`]);
+        let codigo = this.activatedRoute.snapshot.paramMap.get('veiculoId');
+        this.router.navigate([`veiculo/${codigo}/abastecimentos`]);
     }
     this.veiculos = this.veiculoService.selecionarTodos();
   }
@@ -33,11 +30,6 @@ export class VeiculosComponent implements OnInit{
   excluirVeiculo(veiculoId: number){
     this.veiculoService.deletarVeiculo(veiculoId);
     this.router.navigate(['redirectHome']);
-  }
-
-  editarVeiculo(veiculoId:number){
-    let veiculo = this.veiculoService.recuperaVeiculo(veiculoId);
-    this.router.navigate(['veiculos/cadastrar']);
   }
 
 }
