--- src/app/errors/not-found/not-found.component.html
@@ -1,7 +1,0 @@
-<div class="text-center">
-  <h2>This page is not avaliable</h2>
-  <p>
-      The link you have acccessed may be broken or
-      the page may have ben removed.
-  </p>
-</div>
