+++ src/app/model/veiculo.ts
