--- src/app/rodape/rodape.component.css
@@ -1,5 +1,0 @@
-.rodape__link
-{
-  color: white
-}
-
