+++ src/app/service/veiculo.service.ts
@@ -1,15 +1,11 @@
-import { ActivatedRoute, Router } from '@angular/router';
-import { WebStorageUtil } from './../util/web-storage-util';
-import { Observable, of } from 'rxjs';
+import { Observable } from 'rxjs';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { VeiculoStorage } from './veiculo.storage';
-
-import { Veiculo } from './veiculo';
 import { Injectable } from '@angular/core';
 
-const API = 'http://localhost:3000/veiculos';
+import { Veiculo } from '../model/veiculo';
 
-const CHAVE_STORAGE = 'VEICULOS';
+
+const API = 'http://localhost:3000/veiculos';
 
 @Injectable({providedIn: "root"})
 export class VeiculoService
@@ -26,10 +22,7 @@ export class VeiculoService
     headers: new HttpHeaders({'Content-Type': 'application/json'}),
   }
 
-  constructor(public veiculoStorage: VeiculoStorage,
-               private http: HttpClient,
-               private router: Router){
-               //this.veiculos = WebStorageUtil.get(CHAVE_STORAGE);
+  constructor(private http: HttpClient){
   }
 
   cadastrarVeiculo(veiculo: Veiculo){
@@ -38,25 +31,19 @@ export class VeiculoService
 
     this.veiculos.push(veiculo);
 
-    this.insereVeiculo(veiculo);
-
-    WebStorageUtil.set(CHAVE_STORAGE, this.veiculos);
-
-    this.veiculosList = WebStorageUtil.get(CHAVE_STORAGE);
-    console.log(this.veiculosList);
-    console.log("A6");
-
-    this.router.navigate([`/redirectHome`]);
+    this.cadastraVeiculo(veiculo).subscribe(() => alert(`Veiculo ${veiculo.marca} ${veiculo.modelo} cadastrado com sucesso!`));
 
     console.log("veiculoService.cadastrarVeiculo foi finalizada.");
 
   }
 
+  cadastraVeiculo(veiculo: Veiculo): Observable<any> {
+    return this.http.post<any>(API, veiculo);
+  }
+
   async insereVeiculo(veiculo: Veiculo){
 
       console.log("veiculoService.insereVeiculo está sendo executado.");
-
-      console.log(this.veiculosList);
 
       await new Promise<void>((resolve, reject) => {
 
@@ -79,7 +66,7 @@ export class VeiculoService
 
     this.veiculos = this.selecionarTodosVeiculos();
 
-    WebStorageUtil.set(CHAVE_STORAGE, this.veiculos);
+    //this.getVeiculos().subscribe(veics => this.veiculos = veics);
 
     console.log("veiculoService.selecionarTodos foi finalizada.");
 
@@ -122,46 +109,10 @@ export class VeiculoService
     return listVeics;
   }
 
-   async selecionarVeiculo(idVeiculo:number) : Promise<Veiculo>{
-
-    console.log("veiculoService.selecionarVeiculo está sendo executado.");
-
-    console.log("S");
-    console.log(this.veiculos);
-    console.log("U");
-
-    await new Promise<Veiculo>((resolve, reject) => {
-
-      this.http.get<Veiculo>(`${API}/${idVeiculo}`).subscribe({
-         next: (res:any)=>{
-                             const v:Veiculo = new Veiculo(res.id, res.marca, res.modelo, res.anoFabricacao, res.anoModelo, res.odometro, res.litrosTanque, res.combustivel, res.cambio);
-                             this.veiculo = v;
-                             console.log("Promise<Veiculo> this.veiculo");
-                             console.log(this.veiculo);
-                        resolve(this.veiculo);
-               },
-         error: (err:any) =>{
-           reject(console.log(err));
-         }
-       });
-     });
-
-    console.log("veiculoService.selecionarVeiculo foi finalizada.");
-
-    return this.veiculo;
-  }
-
-  selVeic(idVeiculo: number) : Observable<Veiculo>{
-    console.log("veiculoService getVeiculos foi chamado");
-
-    return this.http.get<Veiculo>(`${API}/${idVeiculo}`);
-  }
-
   deletarVeiculo(idVeiculo: number){
 
     console.log("veiculoService.deletarVeiculo está sendo executado.");
 
-    //this.veiculoStorage.deletarVeiculo(idVeiculo);
     this.excluirVeiculo(idVeiculo);
 
     console.log("veiculoService.deletarVeiculo foi finalizada.");
@@ -170,8 +121,6 @@ export class VeiculoService
   excluirVeiculo(idVeiculo:number){
 
     console.log("veiculoService.excluirVeiculo está sendo executado.");
-
-    console.log(`${API}/${idVeiculo}`);
 
     new Promise<void>((resolve, reject) => {
 
@@ -187,20 +136,6 @@ export class VeiculoService
 
     console.log("veiculoService.excluirVeiculo foi finalizada.");
 
-
   }
-
-recuperaVeiculo(idVeiculo: number){
-
-  console.log("veiculoService.recuperVeiculo está sendo executado");
-
-  let veiculo = this.veiculos.filter(v => v.id == idVeiculo);
-
-  console.log(veiculo);
-
-  return veiculo;
-
-}
-
 
 }
