+++ src/index.html
@@ -9,7 +9,7 @@
   <link rel="icon" type="image/x-icon" href="favicon.ico">
   <meta name="author" content="Rafael Moreira">
 	<meta name="keywords" content="Abastecimento, Controle de Abastecimento, Controle Veicular, Combustível">
-	<meta name="description" content="Este app deve ajudar o usuário a controlar os abastecimentos do veículo, valores gastos, postos, combustíveis utilizados e rendimento.">
+	<meta name="description" content="Este app deve ajudar o usuário a controlar os abastecimentos do veículo, valores gastos, postos e combustíveis utilizados.">
 	<meta name="robots" content="all">
 </head>
 <body>
